Add/restore support for custom resources in PointsToGrid
* Separate data from algorithm

* Add custom resource support

* Add docs

* Add support for cudax::async_resource

* Fix default alignment

* Add cleanup

* Split up DeviceResource into separate file

* Fix file
